+++ CHANGELOG.md
@@ -30,3 +30,10 @@
 ## 0.5.12
 
 - **Fix**: Update type of `viewModel` in `ConfigurableController` to `ImplicitlyUwrappedOptional<ViewModelT>` instead of `ViewModelT`
+
+## 0.5.13
+
+- **Change**: Remove type erasure behavior from `AnyBaseTableRow`
+- **Change**: Rename `AnyBaseTableRow` class to `SeparatorRowBox`
+- **Change**: Move `anyRow` property from `EmptyCellRow` to `TableRow` extension and rename it to `separatorRowBox`.
+- **Change**: Move `configure(extreme: middle:)` method from `TableDirector` extension to `Array` extension and rename it to `configureSeparators(extreme: middle:)`